Move speech recognition to transcriber
